fix migration

+++ database/migrations/2025_07_21_135014_create_pembelians_table.php
@@ -11,29 +11,31 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('pembelians', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->char('jenis', 1)->default('B');
-            $table->string('nama_supir', 100)->nullable();
-            $table->string('no_spb', 10)->unique()->nullable();
-            $table->string('nama_barang');
-            $table->string('no_container')->nullable();
-            $table->string('no_container_antar_pulau')->nullable();
-            $table->string('no_surat_jalan')->nullable();
-            $table->string('brondolan')->nullable();
-            $table->string('plat_polisi')->nullable();
-            $table->integer('bruto');
-            $table->integer('tara')->nullable();
-            $table->integer('netto')->nullable();
-            $table->string('keterangan')->nullable();
-            $table->time('jam_masuk');
-            $table->time('jam_keluar')->nullable();
-            $table->timestamps();
-            $table->foreignId('id_supplier')->nullable()->constrained('suppliers')->onDelete('cascade');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->integer('jumlah_karung')->nullable();
-            $table->longText('foto')->nullable(); // For longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_bin DEFAULT NULL
-        });
+        if (!Schema::hasTable('pembelians')) {
+            Schema::create('pembelians', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->char('jenis', 1)->default('B');
+                $table->string('nama_supir', 100)->nullable();
+                $table->string('no_spb', 10)->unique()->nullable();
+                $table->string('nama_barang');
+                $table->string('no_container')->nullable();
+                $table->string('no_container_antar_pulau')->nullable();
+                $table->string('no_surat_jalan')->nullable();
+                $table->string('brondolan')->nullable();
+                $table->string('plat_polisi')->nullable();
+                $table->integer('bruto');
+                $table->integer('tara')->nullable();
+                $table->integer('netto')->nullable();
+                $table->string('keterangan')->nullable();
+                $table->time('jam_masuk');
+                $table->time('jam_keluar')->nullable();
+                $table->timestamps();
+                $table->foreignId('id_supplier')->nullable()->constrained('suppliers')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->integer('jumlah_karung')->nullable();
+                $table->longText('foto')->nullable(); // For longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_bin DEFAULT NULL
+            });
+        }
     }
 
     /**
