+++ database/migrations/2025_07_21_135137_create_laporan_lumbungs_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('laporan_lumbungs', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->char('jenis', 2)->default('IO');
-            $table->string('kode', 10)->unique()->nullable();
-            $table->timestamps();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->string('lumbung', 20)->nullable();
-            $table->string('status_silo', 25)->nullable();
-            $table->boolean('status')->nullable();
-            $table->foreignId('silo_id')->nullable()->constrained('silos')->onDelete('cascade');
-            $table->string('keterangan')->nullable();
-        });
+        if (!Schema::hasTable('laporan_lumbungs')) {
+            Schema::create('laporan_lumbungs', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->char('jenis', 2)->default('IO');
+                $table->string('kode', 10)->unique()->nullable();
+                $table->timestamps();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->string('lumbung', 20)->nullable();
+                $table->string('status_silo', 25)->nullable();
+                $table->boolean('status')->nullable();
+                $table->foreignId('silo_id')->nullable()->constrained('silos')->onDelete('cascade');
+                $table->string('keterangan')->nullable();
+            });
+        }
     }
 
     /**
