+++ database/migrations/2025_07_21_135220_create_dryers_table.php
@@ -11,23 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('dryers', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->char('jenis', 1)->default('D');
-            $table->string('no_dryer', 10)->unique()->nullable();
-            $table->foreignId('id_kapasitas_dryer')->constrained('kapasitas_dryers')->onDelete('cascade');
-            $table->string('operator')->nullable();
-            $table->string('nama_barang')->nullable();
-            $table->double('rencana_kadar')->nullable();
-            $table->double('hasil_kadar')->nullable();
-            $table->integer('total_netto')->nullable();
-            $table->string('pj', 50)->nullable();
-            $table->enum('status', ['processing', 'completed'])->default('processing');
-            $table->timestamps();
-            $table->string('no_cc', 50)->nullable();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('laporan_lumbung_id')->nullable()->constrained('laporan_lumbungs')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('dryers')) {
+            Schema::create('dryers', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->char('jenis', 1)->default('D');
+                $table->string('no_dryer', 10)->unique()->nullable();
+                $table->foreignId('id_kapasitas_dryer')->constrained('kapasitas_dryers')->onDelete('cascade');
+                $table->string('operator')->nullable();
+                $table->string('nama_barang')->nullable();
+                $table->double('rencana_kadar')->nullable();
+                $table->double('hasil_kadar')->nullable();
+                $table->integer('total_netto')->nullable();
+                $table->string('pj', 50)->nullable();
+                $table->enum('status', ['processing', 'completed'])->default('processing');
+                $table->timestamps();
+                $table->string('no_cc', 50)->nullable();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('laporan_lumbung_id')->nullable()->constrained('laporan_lumbungs')->onDelete('cascade');
+            });
+        }
     }
 
     /**
