+++ database/migrations/2025_07_21_135548_create_penjualan_antar_pulaus_table.php
@@ -11,21 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('penjualan_antar_pulaus', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->char('jenis', 2)->default('CJ');
-            $table->string('kode', 10)->unique()->nullable();
-            $table->string('nama_barang');
-            $table->string('no_container')->nullable();
-            $table->integer('netto');
-            $table->integer('netto_diterima')->nullable();
-            $table->string('nama_ekspedisi')->nullable();
-            $table->string('kode_segel')->nullable();
-            $table->timestamps();
-            $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
-            $table->foreignId('kapasitas_kontrak_jual_id')->nullable()->constrained('kapasitas_kontrak_juals')->onDelete('cascade');
-            $table->string('status', 25)->nullable();
-        });
+        if (!Schema::hasTable('penjualan_antar_pulaus')) {
+            Schema::create('penjualan_antar_pulaus', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->char('jenis', 2)->default('CJ');
+                $table->string('kode', 10)->unique()->nullable();
+                $table->string('nama_barang');
+                $table->string('no_container')->nullable();
+                $table->integer('netto');
+                $table->integer('netto_diterima')->nullable();
+                $table->string('nama_ekspedisi')->nullable();
+                $table->string('kode_segel')->nullable();
+                $table->timestamps();
+                $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
+                $table->foreignId('kapasitas_kontrak_jual_id')->nullable()->constrained('kapasitas_kontrak_juals')->onDelete('cascade');
+                $table->foreignId('pembelian_antar_pulau_id')->nullable()->constrained('pembelian_antar_pulaus')->onDelete('cascade');
+                $table->string('status', 25)->nullable();
+            });
+        }
     }
 
     /**
