+++ database/migrations/2025_07_21_135722_create_alamat_kontraks_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('alamat_kontraks', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->foreignId('id_kontrak')->constrained('kontraks')->onDelete('cascade');
-            $table->string('alamat');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('alamat_kontraks')) {
+            Schema::create('alamat_kontraks', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->foreignId('id_kontrak')->constrained('kontraks')->onDelete('cascade');
+                $table->string('alamat');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
