+++ database/migrations/2025_07_21_135751_create_kendaraan_masuks_table.php
@@ -11,23 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('kendaraan_masuks', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('status');
-            $table->string('nama_sup_per');
-            $table->string('plat_polisi')->nullable();
-            $table->string('nama_barang')->nullable();
-            $table->string('keterangan')->nullable();
-            $table->time('jam_masuk')->nullable();
-            $table->time('jam_keluar')->nullable();
-            $table->timestamps();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->integer('nomor_antrian')->nullable();
-            $table->integer('status_selesai')->nullable();
-            $table->integer('status_awal')->nullable();
-            $table->longText('foto')->nullable();
-            $table->string('jenis', 20)->nullable();
-        });
+        if (!Schema::hasTable('kendaraan_masuks')) {
+            Schema::create('kendaraan_masuks', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('status');
+                $table->string('nama_sup_per');
+                $table->string('plat_polisi')->nullable();
+                $table->string('nama_barang')->nullable();
+                $table->string('keterangan')->nullable();
+                $table->time('jam_masuk')->nullable();
+                $table->time('jam_keluar')->nullable();
+                $table->timestamps();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->integer('nomor_antrian')->nullable();
+                $table->integer('status_selesai')->nullable();
+                $table->integer('status_awal')->nullable();
+                $table->longText('foto')->nullable();
+                $table->string('jenis', 20)->nullable();
+            });
+        }
     }
 
     /**
