+++ database/migrations/2025_07_21_135808_create_kendaraan_muats_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('kendaraan_muats', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('nama_supir');
-            $table->foreignId('kendaraan_id')->constrained('kendaraans')->onDelete('cascade');
-            $table->integer('tonase');
-            $table->string('tujuan');
-            $table->timestamp('jam_masuk')->nullable();
-            $table->timestamp('jam_keluar')->nullable();
-            $table->integer('status')->nullable();
-            $table->integer('status_awal')->nullable();
-            $table->timestamps();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->longText('foto')->nullable();
-            $table->string('keterangan')->nullable();
-        });
+        if (!Schema::hasTable('kendaraan_muats')) {
+            Schema::create('kendaraan_muats', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('nama_supir');
+                $table->foreignId('kendaraan_id')->constrained('kendaraans')->onDelete('cascade');
+                $table->integer('tonase');
+                $table->string('tujuan');
+                $table->timestamp('jam_masuk')->nullable();
+                $table->timestamp('jam_keluar')->nullable();
+                $table->integer('status')->nullable();
+                $table->integer('status_awal')->nullable();
+                $table->timestamps();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->longText('foto')->nullable();
+                $table->string('keterangan')->nullable();
+            });
+        }
     }
 
     /**
