+++ database/migrations/2025_07_22_155753_create_pembelians_records_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('pembelian_records', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->foreignId('pembelian_id')->constrained('pembelians')->onDelete('cascade');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('pembelian_records')) {
+            Schema::create('pembelian_records', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->foreignId('pembelian_id')->constrained('pembelians')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
